Move SubmitModal into utils
