+++ src/js/components/utils/useFetchTasksIfMissing.tsx
@@ -25,7 +25,7 @@ export default (
         fetchObjects({
           objectType: OBJECT_TYPES.TASK,
           filters: { project: project.id },
-          shouldSubscribeToObject: true, 
+          shouldSubscribeToObject: true,
         }),
       );
     }
